+++ lib/features/home/view/home_screen.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:get_it/get_it.dart';
-import 'package:hive_flutter/hive_flutter.dart';
-
-import '../../../models/task/task.dart';
+import 'package:pixel_todo/features/home/widgets/widgets.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
@@ -40,102 +37,9 @@ class HomeScreen extends StatelessWidget {
               ),
             ),
             const Expanded(
-                flex: 4, child: CustomScrollView(slivers: [SliverTasks()]))
+                flex: 4, child: CustomScrollView(slivers: [TasksList()]))
           ],
         ),
-      ),
-    );
-  }
-}
-
-class SliverTasks extends StatefulWidget {
-  const SliverTasks({
-    super.key,
-  });
-
-  @override
-  State<SliverTasks> createState() => _SliverTasksState();
-}
-
-class _SliverTasksState extends State<SliverTasks> {
-  final Box<Task> taskBox = GetIt.I<Box<Task>>();
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder(
-      valueListenable: taskBox.listenable(),
-      builder: (context, Box<Task> box, _) {
-        if (box.isEmpty) {
-          return const SliverToBoxAdapter(
-            child: Center(child: Text('Нет задач')),
-          );
-        }
-        return SliverList(
-          delegate: SliverChildBuilderDelegate(
-            (context, index) {
-              final task = box.getAt(index);
-              return task != null
-                  ? TaskItem(task: task, index: index)
-                  : Container();
-            },
-            childCount: box.length,
-          ),
-        );
-      },
-    );
-  }
-}
-
-class TaskItem extends StatelessWidget {
-  final Task task;
-  final int index;
-
-  const TaskItem({super.key, required this.task, required this.index});
-
-  @override
-  Widget build(BuildContext context) {
-    final Box<Task> taskBox = Hive.box<Task>('todo_box_name');
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-      child: Stack(
-        alignment: Alignment.center,
-        children: [
-          Image.asset(
-            'assets/images/pergament.png',
-            width: double.infinity,
-            height: 100,
-            fit: BoxFit.fill,
-          ),
-          Positioned(
-            left: 16,
-            child: Checkbox(
-              activeColor: Colors.red,
-              checkColor: Colors.blue,
-              value: task.isCompleted,
-              onChanged: (value) {
-                taskBox.putAt(
-                    index,
-                    Task(
-                        taskTitle: task.taskTitle,
-                        isCompleted: value ?? false));
-              },
-            ),
-          ),
-          Positioned(
-            left: 60,
-            right: 16, // Добавляем right для ограничения ширины
-            child: Text(
-              task.taskTitle,
-              maxLines: 2, // Максимум 2 строки
-              overflow: TextOverflow.ellipsis, // Многоточие, если не помещается
-              softWrap: true, // Перенос текста на новую строку
-              style: const TextStyle(
-                  color: Colors.black,
-                  fontSize: 18,
-                  fontFamily: "TeletactileRus"),
-            ),
-          ),
-        ],
       ),
     );
   }
